journalrecord uint64

+++ Main/JournalRecord.hpp
@@ -8,19 +8,19 @@
 class JournalRecord {
 
 private:
-    uint32_t transaction_id;
-    DArray<uint32_t> *values;
+    uint64_t transaction_id;
+    DArray<uint64_t> *values;
 	bool type;
 
 public:
-    JournalRecord(uint32_t, bool type);
+    JournalRecord(uint64_t, bool type);
 	~JournalRecord();
 
 	bool getType();
-	uint32_t getValue(uint32_t);
-	uint32_t getTransactionId();
-	DArray<uint32_t> *getAllValues();
+	uint64_t getValue(uint64_t);
+	uint64_t getTransactionId();
+	DArray<uint64_t> *getAllValues();
 
-	void addValue(uint32_t);
+	void addValue(uint64_t);
 	void printRecord();
 };
